Move tests around, temporarily disable `std-checks`
