+++ tests/perf/function-contract/const-fn-with-effect/Cargo.toml
@@ -1,0 +1,14 @@
+# Copyright Kani Contributors
+# SPDX-License-Identifier: Apache-2.0 OR MIT
+
+[package]
+name = "const-fn-with-effect"
+version = "0.1.0"
+edition = "2021"
+
+# See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
+
+[dependencies]
+
+[package.metadata.kani]
+unstable = { function-contracts = true, mem-predicates = true, ghost-state = true, uninit-checks = true }
