+++ tests/perf/function-contract/const-fn-with-effect/expected
