+++ tests/perf/function-contract/const-fn/expected
