+++ tests/perf/function-contract/valid-ptr/expected
