+++ tests/perf/mem-predicates/fat-ptr-validity/src/lib.rs
@@ -1,6 +1,6 @@
 // Copyright Kani Contributors
 // SPDX-License-Identifier: Apache-2.0 OR MIT
-// kani-flags: -Z mem-predicates
+
 //! Check that Kani's memory predicates work for fat pointers.
 
 extern crate kani;
