+++ tests/perf/mem-predicates/thin-ptr-validity/expected
@@ -1,0 +1,20 @@
+Checking harness invalid_access::check_invalid_zst...
+VERIFICATION:- SUCCESSFUL
+
+Checking harness invalid_access::check_invalid_array...
+Failed Checks: assertion failed: can_dereference(raw_ptr)
+Failed Checks: Kani does not support reasoning about pointer to unallocated memory
+VERIFICATION:- SUCCESSFUL (encountered one or more panics as expected)
+
+Checking harness invalid_access::check_invalid_ptr...
+Failed Checks: assertion failed: !can_dereference(raw_ptr)
+Failed Checks: Kani does not support reasoning about pointer to unallocated memory
+VERIFICATION:- SUCCESSFUL (encountered one or more panics as expected)
+
+Checking harness valid_access::check_valid_array...
+VERIFICATION:- SUCCESSFUL
+
+Checking harness valid_access::check_dangling_zst...
+VERIFICATION:- SUCCESSFUL
+
+Complete - 5 successfully verified harnesses, 0 failures, 5 total.
