+++ tests/perf/uninit/expected
@@ -1,0 +1,1 @@
+Complete - 5 successfully verified harnesses, 0 failures, 5 total.
